feat(permission): enhance permission management with new permissions and authorization checks

+++ src/auth/router/tests/test_endpoint_permission.py
@@ -1,7 +1,8 @@
-def test_create_permission(client):
+def test_create_permission(client, create_user_with_group, create_token_admin):
     response = client.post(
         "/permissions/",
-        json={"name": "read_articles", "description": "Permission to read articles"}
+        json={"name": "read_articles", "description": "Permission to read articles"},
+        headers={"Authorization": f"Bearer {create_token_admin}"}
     )
     assert response.status_code == 200
     assert response.json() == {
@@ -9,43 +10,45 @@ def test_create_permission(client):
         'description': 'Permission to read articles'
     }
 
-def test_create_when_exists_permission(client, create_permission):
+def test_create_when_exists_permission(client, create_permission, create_user_with_group, create_token_admin):
     response = client.post(
         "/permissions/",
-        json={"name": "testpermission", "description": "A test permission"}
+        json={"name": "testpermission", "description": "A test permission"},
+        headers={"Authorization": f"Bearer {create_token_admin}"}
     )
     assert response.status_code == 400
     assert response.json() == {
         "detail": "Permission with this name already exists."
     }
 
-def test_delete_permission(client, create_permission):
-    response = client.delete("/permissions/1")
+def test_delete_permission(client, create_permission, create_user_with_group, create_token_admin):
+    response = client.delete("/permissions/1", headers={"Authorization": f"Bearer {create_token_admin}"})
     assert response.status_code == 200
     assert response.json() == {"detail": "Permission deleted successfully."}
 
-def test_delete_nonexistent_permission(client):
-    response = client.delete("/permissions/999")
+def test_delete_nonexistent_permission(client, create_permission, create_user_with_group, create_token_admin):
+    response = client.delete("/permissions/999", headers={"Authorization": f"Bearer {create_token_admin}"})
     assert response.status_code == 404
     assert response.json() == {"detail": "Permission not found."}
 
-def test_get_permission_by_id(client, create_permission):
-    response = client.get("/permissions/1")
+def test_get_permission_by_id(client, create_permission, create_user_with_group, create_token_admin):
+    response = client.get("/permissions/1", headers={"Authorization": f"Bearer {create_token_admin}"})
     assert response.status_code == 200
     assert response.json() == {
         'name': 'testpermission', 
         'description': 'A test permission'
     }
 
-def test_get_nonexistent_permission_by_id(client):
-    response = client.get("/permissions/999")
+def test_get_nonexistent_permission_by_id(client, create_permission, create_user_with_group, create_token_admin):
+    response = client.get("/permissions/999", headers={"Authorization": f"Bearer {create_token_admin}"})
     assert response.status_code == 404
     assert response.json() == {"detail": "Permission not found."}
 
-def test_update_permission(client, create_permission):
+def test_update_permission(client, create_permission, create_user_with_group, create_token_admin):
     response = client.put(
         "/permissions/1",
-        json={"name": "updated_permission", "description": "Updated description"}
+        json={"name": "updated_permission", "description": "Updated description"},
+        headers={"Authorization": f"Bearer {create_token_admin}"}
     )
     assert response.status_code == 200
     assert response.json() == {
@@ -53,10 +56,11 @@ def test_update_permission(client, create_permission):
         'description': 'Updated description'
     }
 
-def test_update_nonexistent_permission(client):
+def test_update_nonexistent_permission(client, create_permission, create_user_with_group, create_token_admin):
     response = client.put(
         "/permissions/999",
-        json={"name": "nonexistent_permission", "description": "Should not exist"}
+        json={"name": "nonexistent_permission", "description": "Should not exist"},
+        headers={"Authorization": f"Bearer {create_token_admin}"}
     )
     assert response.status_code == 404
     assert response.json() == {"detail": "Permission not found."}
